Fix #9 and move source files into src directory
