+++ src/alltypes.d.ts
