+++ src/components/footer/index.tsx
@@ -1,5 +1,5 @@
 import Linker, { LinkerProps } from './linker'
-import styles from '/styles/Linker.module.css'
+import styles from '@src/styles/Linker.module.css'
 import classNames from 'classnames/bind'
 
 const cx = classNames.bind(styles)
